+++ apps/lfx-one/src/app/modules/crowdfunding/initiative-detail/components/initiative-settings-drawer/components/settings-branding-tab/settings-branding-tab.component.html
@@ -1,0 +1,89 @@
+<!-- Copyright The Linux Foundation and each contributor to LFX. -->
+<!-- SPDX-License-Identifier: MIT -->
+
+<div class="flex flex-col gap-5" data-testid="settings-branding-tab">
+  <h3 class="text-xs font-bold text-gray-500 uppercase tracking-wide">Branding</h3>
+
+  <div class="flex flex-col gap-1.5" data-testid="settings-field-logo">
+    <label class="text-sm font-medium text-gray-700"> Initiative logo <span class="text-red-500" aria-hidden="true">*</span> </label>
+
+    <!-- Uploading state -->
+    @if (uploadingLogo()) {
+      <div class="border border-dashed border-gray-300 rounded-xl px-5 py-6 flex items-center gap-5" data-testid="settings-logo-uploading">
+        <div class="w-12 h-12 rounded-full bg-gray-100 border border-gray-200 flex items-center justify-center shrink-0 overflow-hidden relative">
+          @if (logoUrl()) {
+            <img [src]="logoUrl()" alt="Logo preview" class="w-full h-full object-cover opacity-50" />
+          } @else {
+            <i class="fa-light fa-image text-gray-400 text-xl" aria-hidden="true"></i>
+          }
+          <div class="absolute inset-0 flex items-center justify-center">
+            <i class="fa-solid fa-spinner-third text-gray-500 animate-spin" aria-hidden="true"></i>
+          </div>
+        </div>
+        <span class="text-sm text-gray-500">Uploading…</span>
+      </div>
+    }
+
+    <!-- Has logo state -->
+    @else if (logoUrl()) {
+      <div class="flex items-center gap-4 p-4 border border-gray-200 rounded-lg bg-gray-50" data-testid="settings-logo-preview-container">
+        <div
+          class="w-16 h-16 rounded-lg border border-gray-200 bg-white flex items-center justify-center overflow-hidden flex-shrink-0"
+          data-testid="settings-logo-preview"
+          aria-label="Current initiative logo">
+          <img [src]="logoUrl()" [alt]="initiative().name + ' logo'" class="w-full h-full object-contain p-1" />
+        </div>
+        <div class="flex flex-col gap-1.5">
+          <p class="text-xs text-gray-500">PNG, JPG, GIF, or WebP · Max 2 MB</p>
+          <div class="flex items-center gap-2 mt-1">
+            <lfx-button
+              label="Remove"
+              icon="fa-regular fa-trash-can"
+              severity="secondary"
+              size="small"
+              (onClick)="removeLogo()"
+              data-testid="settings-logo-remove-btn" />
+            <lfx-button
+              label="Choose different"
+              icon="fa-solid fa-upload"
+              severity="secondary"
+              size="small"
+              (onClick)="triggerLogoUpload()"
+              data-testid="settings-logo-upload-btn" />
+          </div>
+        </div>
+      </div>
+    }
+
+    <!-- Empty state -->
+    @else {
+      <div
+        class="border border-dashed border-gray-300 rounded-xl px-5 py-6 flex items-center justify-center gap-5 cursor-pointer hover:bg-gray-50 transition-colors"
+        (click)="triggerLogoUpload()"
+        role="button"
+        tabindex="0"
+        (keydown.enter)="$event.preventDefault(); triggerLogoUpload()"
+        (keydown.space)="$event.preventDefault(); triggerLogoUpload()"
+        data-testid="settings-logo-empty">
+        <i class="fa-solid fa-image text-gray-200 text-4xl shrink-0" aria-hidden="true"></i>
+        <div class="flex flex-col gap-2">
+          <p class="text-sm text-gray-900 leading-5"><span class="text-blue-600">Click to upload</span></p>
+          <p class="text-xs text-gray-400 leading-4">PNG, JPG, GIF, or WebP · Max 2 MB · 600×600px</p>
+        </div>
+      </div>
+    }
+
+    @if (logoUploadError()) {
+      <p class="text-xs text-red-500" role="alert" data-testid="settings-logo-error">{{ logoUploadError() }}</p>
+    }
+
+    <input
+      #logoFileInput
+      type="file"
+      accept="image/png,image/jpeg,image/gif,image/webp"
+      class="hidden"
+      aria-hidden="true"
+      (change)="onLogoFileSelected($event)"
+      data-testid="settings-logo-file-input" />
+  </div>
+</div>
