+++ apps/lfx-one/src/app/modules/crowdfunding/initiative-detail/components/initiative-settings-drawer/components/settings-funding-tab/settings-funding-tab.component.html
@@ -1,0 +1,81 @@
+<!-- Copyright The Linux Foundation and each contributor to LFX. -->
+<!-- SPDX-License-Identifier: MIT -->
+
+<div class="flex flex-col gap-5" data-testid="settings-funding-tab">
+  <h3 class="text-xs font-bold text-gray-500 uppercase tracking-wide">Funding</h3>
+
+  <p class="text-sm text-gray-500">
+    Provide your initial estimated general fund funding goal. You can update your goal at any time and continue raising funds after your goal is met.
+  </p>
+
+  <!-- Funding Goal -->
+  <div class="flex flex-col gap-1.5 max-w-[240px]" data-testid="settings-field-goal">
+    <label for="settings-goal" class="text-sm font-medium text-gray-700"> Funding Goal <span class="text-red-500" aria-hidden="true">*</span> </label>
+    <lfx-input-text [form]="form" control="goal" id="settings-goal" type="number" placeholder="0" icon="fa-light fa-dollar-sign" />
+  </div>
+
+  <!-- Distribution header -->
+  <div class="flex flex-col gap-1" data-testid="settings-distribution-header">
+    <p class="text-sm font-medium text-gray-700">Fund Distribution</p>
+    <p class="text-xs text-gray-500 leading-4">Allocate your funds across categories. Enabled categories must have a percentage greater than zero.</p>
+  </div>
+
+  <!-- Progress bar -->
+  @if (hasEnabledCategories()) {
+    <div class="flex flex-col gap-1" data-testid="settings-distribution-progress">
+      <div class="flex items-center gap-2">
+        <div class="flex-1 h-1.5 rounded-full bg-gray-200 overflow-hidden">
+          <div
+            class="h-full rounded-full transition-all"
+            [class]="totalAllocated() > 100 ? 'bg-red-500' : 'bg-amber-500'"
+            [style.width.%]="totalAllocated() > 100 ? 100 : totalAllocated()"></div>
+        </div>
+        <span class="text-xs font-semibold text-gray-900 shrink-0">{{ totalAllocated() }}%</span>
+      </div>
+      <p class="text-xs leading-4" [class]="totalAllocated() > 100 ? 'text-red-600' : 'text-amber-600'" data-testid="settings-distribution-remaining">
+        @if (remaining() >= 0) {
+          {{ remaining() }}% remaining
+        } @else {
+          {{ -remaining() }}% over budget
+        }
+      </p>
+    </div>
+  }
+
+  <!-- Distribution categories -->
+  <div class="flex flex-col" data-testid="settings-distribution-categories">
+    @for (item of distributionItems(); track item.category; let i = $index; let last = $last) {
+      <div
+        class="flex items-center gap-4 py-4"
+        [class.border-b]="!last"
+        [class.border-gray-200]="!last"
+        [attr.data-testid]="'settings-category-' + item.category">
+        <p-toggleSwitch [ngModel]="item.enabled" (onChange)="toggleCategory(i, $event.checked)" [ariaLabel]="'Enable ' + item.label + ' category'" />
+        <div class="flex-1 flex flex-col gap-0.5 min-w-0">
+          <p class="text-sm text-gray-900">{{ item.label }}</p>
+          <p class="text-xs text-gray-500 leading-4">{{ item.description }}</p>
+        </div>
+        @if (item.enabled) {
+          <div class="flex items-center gap-1.5 shrink-0" [attr.data-testid]="'settings-category-enabled-' + item.category">
+            <input
+              type="number"
+              min="0"
+              max="100"
+              [value]="item.percentage > 0 ? item.percentage : ''"
+              placeholder="0"
+              (input)="updatePercentage(i, $event)"
+              class="w-14 px-2 py-1.5 text-sm text-center border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+              [attr.aria-label]="item.label + ' percentage'"
+              [attr.data-testid]="'settings-category-pct-' + item.category" />
+            <span class="text-xs text-gray-500">%</span>
+            <span class="text-sm text-gray-900 w-10 text-right" [attr.data-testid]="'settings-category-amount-' + item.category">
+              {{ distributionAmounts()[i] }}
+            </span>
+          </div>
+        } @else {
+          <div class="w-32 shrink-0"></div>
+        }
+      </div>
+    }
+  </div>
+</div>
